+++ code_backend/organize_playlist.py
@@ -1,13 +1,11 @@
-from spotify_access import spotify_client, request_multiple_items_of_same_type
-from secondary_methods import image_to_b64, url_to_uri, split_list_into_chunks
-from shared_config import *
+from code_backend.shared_config import *
+from code_backend.secondary_methods import image_to_b64, url_to_uri, split_list_into_chunks, uri_to_id, print_error, load_json, value_from_dict, key_from_dict,print_debug
+import code_backend.spotify_web_api as sp_api
 
-# Warning: Not using `spotify_web_api.py` yet
 
-sp = spotify_client()
 DEFAULT_IMAGE = image_to_b64(Image.open(NO_IMAGE_PATH), 'PNG')
 
-
+# mps: 3
 def all_shuffle(collection_tracks: list) -> list:
     """
     Shuffles the given list using the PRNG random.randint(). Every item will occur exactly once.
@@ -23,125 +21,31 @@ def all_shuffle(collection_tracks: list) -> list:
         shuffle_track_ids.pop(random_position)
     return shuffled_collection
 
-
-def prepare_collection_tracks(
-        collection: dict,
-        chunk_size: int = 50,
-        shuffle: bool = False
-) -> (list, dict):
+# mps: 3
+def all_shuffle_dict(collection_tracks: dict) -> dict:
     """
-    Extracts every Track URI from a given Collection instance and prepares them for further API calls.
-    :param collection: Collection instance (e.g. from sp.playlist(), sp.album(), etc.)
-    :param chunk_size: API limit for items per request (https://developer.spotify.com/documentation/web-api/reference/get-playlists-tracks -> limit)
-    :param shuffle: If to shuffle the Collection Tracks
-    :return: List of Track URI chunks and Dict with all Tracks
+    Shuffles the given dict using the PRNG random.randint(). Every item will occur exactly once.
+    :param collection_tracks: dict to be shuffled
+    :return: shuffled dict
     """
-    track_ids = []
-    track_dicts = []
+    shuffled_collection = dict()
+    shuffle_track_uris = list(collection_tracks.keys())
 
-    for current_offset in range(0, collection["tracks"]["total"], chunk_size):
-        try:
-            match collection["type"]:
-                case "album":
-                    items = sp.album_tracks(
-                        album_id=collection["id"],
-                        limit=chunk_size,
-                        offset=current_offset,
-                        market=MARKET
-                    )
-                case "playlist":
-                    items = sp.playlist_items(
-                        playlist_id=collection["id"],
-                        limit=chunk_size,
-                        offset=current_offset,
-                        market=MARKET
-                    )
-                case _:
-                    items = None
-
-        except SpotifyException as e:
-            print(f"\n\x1b[31mAn Error occurred while fetching the Collection {collection["id"]} items\x1b[30m\n{e}")
-            sys.exit(1)
-
-        if items is None:
-            print(f"\n\x1b[31mCould not fetch Collection {collection["id"]} items\x1b[30m\n")
-            sys.exit(1)
-
-        # <--- debugging --->
-        # with open("debugging.json", "w") as f:
-        #     json.dump(items, f)
-        # <--- end debugging --->
-
-        match collection["type"]:
-            case "album":
-                for item in items["items"]:
-                    if item["uri"] in track_ids:
-                        continue
-                    track_ids.append(item["uri"])
-                    track_dicts.append({
-                        "uri": item["uri"],
-                        "name": item["name"],
-                        "artist": item["artists"][0]["name"]
-                    })
-            case "playlist":
-                for item in items["items"]:
-                    if item["track"]["uri"] in track_ids:
-                        continue
-                    track_ids.append(item["track"]["uri"])
-                    track_dicts.append({
-                        "uri": item["track"]["uri"],
-                        "name": item["track"]["name"],
-                        "artist": item["track"]["artists"][0]["name"]
-                    })
-
-    if shuffle:
-        track_ids = all_shuffle(track_ids)
-
-    if len(track_ids) < 1 or len(track_dicts) < 1:
-        print(f"\n\x1b[31mCould not fetch Collection {collection["id"]} items\x1b[30m\n")
-        sys.exit(1)
-
-    return track_ids, track_dicts
+    while len(shuffle_track_uris) > 0:
+        random_position = random.randint(0, len(shuffle_track_uris) - 1)
+        current_track_uri = shuffle_track_uris[random_position]
+        shuffled_collection[current_track_uri] = collection_tracks[current_track_uri]
+        shuffle_track_uris.pop(random_position)
+    return shuffled_collection
 
 
-def fetch_collection_with_tracks(collection_id: str, collection_type: Literal["album", "playlist"], shuffle: bool = False) -> (dict, (list, dict)):
-    """
-    Requests the Playlist Dict using the Spotify API then fetches every track from the playlist
-    :param collection_id: Spotify Collection ID
-    :param collection_type: Spotify Collection Type
-    :param shuffle: If to shuffle the Collection Tracks
-    :return: Triple of (1) Collection JSON, (2) list of Track URIs and (3) Dict with all Tracks JSON
-    """
-    try:
-        match collection_type:
-            case "album":
-                collection = sp.album(album_id=collection_id, market=MARKET)
-            case "playlist":
-                collection = sp.playlist(playlist_id=collection_id, market=MARKET)
-            case _:
-                collection = None
-
-    except SpotifyException as e:
-        print(f"\n\x1b[31mAn Error occurred while fetching the Collection {collection_id} items\x1b[30m\n{e}")
-        sys.exit(1)
-
-    if not collection:
-        print(f"\n\x1b[31mCollection with the ID {collection_id} does not exist\x1b[30m\n")
-        sys.exit(1)
-
-    collection_tracks = prepare_collection_tracks(collection=collection, shuffle=shuffle)
-    if collection_tracks is None:
-        print(f"\n\x1b[31mCould not prepare Collection {collection_id} Tracks\x1b[30m\n")
-        sys.exit(1)
-
-    return collection, collection_tracks
-
-
-def remove_duplicates(tracks: list[dict]) -> list:
+# mps: 3
+def remove_duplicates(tracks: list[dict], get_removed: bool = False) -> set | tuple[set, set]:
     """
     Removes duplicate tracks (name and artist are the same) from a list of dicts
     :param tracks: list of tracks in the form of {"uri": ..., "name": ..., "artist": ...}
-    :return: List of unique Tracks
+    :param get_removed: True: (no duplicates, removed duplicates); False: no duplicates
+    :return: Set of unique Tracks (optional with the removed duplicates)
     """
     uniques = {}
     for i in tracks:
@@ -149,28 +53,14 @@ def remove_duplicates(tracks: list[dict]) -> list:
         # Hash produces unique identifiers, if a collision occurs, the URI gets overwritten by the new one -> no collisions/duplicates
         uniques[sha256(a.encode("utf-8")).hexdigest()] = i["uri"]
 
-    return list(uniques.values())
+    all_tracks = set([i["uri"] for i in tracks])
+    keep_tracks = set(uniques.values())
+    remove_tracks = all_tracks - keep_tracks
+
+    return (keep_tracks, remove_tracks) if get_removed else keep_tracks
 
 
-def empty_playlist(collection_id, collection_tracks:list[str]) -> None:
-    """
-    Removing Tracks from Playlist while avoiding API limitations of removing more than 100 items per request
-    :param collection_id: Collection ID
-    :param collection_tracks: List containing Track URIs
-    :return: empty Spotify Collection
-    """
-    limit = 100
-    for current_offset in range(0, len(collection_tracks), limit):
-        try:
-            sp.playlist_remove_all_occurrences_of_items(
-                playlist_id=collection_id,
-                items=collection_tracks[current_offset:current_offset + limit],
-            )
-        except SpotifyException as e:
-            print(f"\n\x1b[31mError occurred while removing Tracks from Playlist {collection_id}\n{e}\x1b[30m")
-            sys.exit(1)
-
-
+# mps: 3
 def organize_collection(collection_uri: str, **kwargs) -> None:
     """
     Reorganizes the given Playlist and either creates new Playlist or updates existing Playlist. (If successful either new Playlist appears in Spotify or Playlist Content changed)
@@ -178,167 +68,109 @@ def organize_collection(collection_uri: str, **kwargs) -> None:
     :return: New/Updated Playlist in Spotify
     """
 
-    _, collection_type, collection_id = collection_uri.split(":")
+    collection_id, collection_type = uri_to_id(spotify_uri=collection_uri, get_type=True)
     if not collection_type in ["album", "playlist"]:
-        print(f"\n\x1b[31m Collection {collection_type} is not valid.\x1b[30m\nValid options are: album, playlist")
-        sys.exit(1)
+        print_error(
+            error_message="Collection {collection_type} is not valid.",
+            more_infos="Valid options are: album, playlist",
+            exit_code=1
+        )
 
     # Process Keyword Arguments
-    shuffle_tracks: bool = ("shuffle" in kwargs and kwargs["shuffle"])  # shuffle exists and is True
-    remove_tracks: list[str] = kwargs["remove"] if "remove" in kwargs else []
+    shuffle_tracks = bool("shuffle" in kwargs and kwargs["shuffle"])  # shuffle exists and is True
+    remove_tracks = list(kwargs["remove"] if "remove" in kwargs else [])
 
-    # Fetch Collection
-    # noinspection PyTypeChecker
-    fetched_collection = fetch_collection_with_tracks(
-        collection_id=collection_id,
-        collection_type=collection_type,
-        shuffle=shuffle_tracks
-    )
+    # Fetch tracks from collection
+    match collection_type:
+        case "album":
+            collection_tracks = sp_api.get_album_tracks(album_id=collection_id)
+        case "playlist":
+            collection_tracks = sp_api.get_playlist_items(playlist_id=collection_id)
+        case _:
+            print(f"{CRED}Invalid collection type: {collection_type}{TEXTCOLOR}")
+            return None
 
-    collection = fetched_collection[0]
-    collection_tracks, track_dict = fetched_collection[1]
+    if len(collection_tracks) < 1:
+        print_error(
+            error_message="Could not fetch items of '{collection_uri}'",
+            exit_code=1
+        )
 
-    # <--- debugging --->
-    # with open("debugging.json", "w") as file:
-    #     json.dump(track_dict, file)
-    #     sys.exit(13)
-    # <--- end debugging --->
+    # get relevant Collection Data
+    match collection_type:
+        case "album":
+            collection = sp_api.get_album(album_id=collection_id)
+            collection_owner_id = [current_artist['id'] for current_artist in collection['artists']][0]
+            collection_owner_name = collection["artists"][0]["name"]
+            collection_owner_type = "Artist"
+        case "playlist":
+            collection = sp_api.get_playlist(playlist_id=collection_id)
+            collection_owner_id = collection[collection_uri]['owner']['id']
+            collection_owner_name = collection[collection_uri]['owner']['display_name']
+            collection_owner_type = "User"
+            old_description = collection[collection_uri]['description']
 
-    # Create Playlist if not yours
-    collection_owner = \
-        collection["owner"]["display_name"] if collection_type == "playlist" \
-        else collection["artists"][0]["name"]
+    # create Playlist if not yours
+    current_user_id = value_from_dict(sp_api.get_current_users_profile())['id']
+    if collection_owner_id != current_user_id:
+        new_playlist = sp_api.create_playlist(
+            user_id=current_user_id,
+            name=f"{collection[collection_uri]["name"]} (Shuffled)",
+            public=True,
+            collaborative=False,
+            description=f"reshuffled Playlist of '{collection[collection_uri]["name"]}' created by {collection_owner_type} '{collection_owner_name}'"
+        )
 
-    if collection_owner != sp.current_user()["display_name"]:
-        try:
-            new_playlist_json = sp.user_playlist_create(
-                user=sp.current_user()['id'],
-                name=f"{collection["name"]} (Shuffled)",
-                public=True,
-                collaborative=False,
-                description=f"reshuffled Playlist of {collection["name"]} created by {collection_owner}"
-            )
-        except SpotifyException as e:
-            print(f"\n\x1b[31mError occurred while creating the Playlist {collection_id}\x1b[30m\n",e)
-            sys.exit(1)
+        collection_id = uri_to_id(key_from_dict(new_playlist))
+        sp_api.add_custom_playlist_cover_image(playlist_id=collection_id, b64_image=DEFAULT_IMAGE)
 
-        if new_playlist_json is None:
-            print(f"\n\x1b[31mCould not create Playlist {collection_id}\x1b[30m\n")
-            sys.exit(1)
-
-        collection_id = new_playlist_json['id']
-
-        try:
-            sp.playlist_upload_cover_image(playlist_id=collection_id, image_b64=DEFAULT_IMAGE)
-        except SpotifyException as e:
-            print(f"\n\x1b[33mSetting Playlists {collection_id} Image did not work\n{e}\x1b[30m\n")
-
-        try:
-            sp.playlist_change_details(playlist_id=collection_id,description="Reshuffled Playlist")
-        except SpotifyException as e:
-            print(f"\n\x1b[33mSetting Playlists {collection_id} Description did not work\n{e}\x1b[30m\n")
-
-
-    # empty Playlist if it is yours
+    # update Playlist name if it is yours (optional step)
     else:
-        empty_playlist(collection_id, collection_tracks)
+        sp_api.change_playlist_details(
+            playlist_id=collection_id,
+            name=f"{collection[collection_uri]["name"]} (Shuffled)",
+            public=True,
+            collaborative=False,
+            description=old_description
+        )
 
-    tracks_to_keep: list[str] = []
-    unwanted_tracks: list[str] = []
+    # Shuffle Tracks if wanted (get RNG on what duplicated tracks to remove)
+    if shuffle_tracks:
+        collection_tracks = all_shuffle_dict(collection_tracks)
 
     # Remove unwanted Tracks
+    tracks_to_keep = set(collection_tracks.keys())
     while len(remove_tracks) > 0:
-        # Note: Can remove Tracks "unintentionally" e.g. if pattern is "Live in" (Live/Concert Version), "Live in the moment" by Portugal. The Man gets removed
         if remove_tracks[0] == "duplicate":
-            # Todo: Check if collection_tracks is the correct attribute
-            print(collection_tracks)
-            unwanted_tracks.extend(remove_duplicates(track_dict))  # Todo continue here
-            print(collection_tracks)
+            # Two Tracks are duplicates, if both track name and artist ID match
+            hashable_collection_tracks = [{"uri": current_track_uri ,"name": current_track["name"], "artist": current_track["artists"][0]["id"]} for current_track_uri, current_track in collection_tracks.items()]
+            tracks_to_keep = remove_duplicates(tracks=hashable_collection_tracks)
 
         else:
-            tmp = remove_tracks[0]
-            for track in track_dict:
-                if not re.search(remove_tracks[0], track["name"], flags=re.IGNORECASE):
-                    if track["uri"] not in tracks_to_keep:
-                        tracks_to_keep.append(track["uri"])
-                else:
-                    if track["uri"] not in tracks_to_keep:
-                        unwanted_tracks.append(track["uri"])
+            debug_1 = remove_tracks[0]
+
+            for current_track_uri, current_track in collection_tracks.items():
+                # remove track if pattern matches
+                # Note: Can remove Tracks "unintentionally" e.g. if pattern is "Live in" (Live/Concert Version), "Live in the moment" by 'Portugal. The Man' gets removed
+                if re.search(pattern=remove_tracks[0], string=current_track["name"], flags=re.IGNORECASE):
+                    tracks_to_keep.remove(current_track_uri)
 
         remove_tracks.pop(0)
 
     if tracks_to_keep:
-        collection_tracks = tracks_to_keep
+        collection_tracks = list(tracks_to_keep)
 
-    split_data = split_list_into_chunks(
-        lst=collection_tracks,
-        # API limit for items per request is 100 (https://developer.spotify.com/documentation/web-api/reference/add-tracks-to-playlist -> uris)
-        chunk_length=100
-    )
+    # Shuffle Tracks if wanted (Shuffle again, to avoid "bias" after removing multiple tracks)
+    if shuffle_tracks:
+        collection_tracks = all_shuffle(list(collection_tracks))
 
-    # add shuffled Tracks to (new) Playlist
-    for chunk in split_data:
-        sp.playlist_add_items(playlist_id=collection_id, items=chunk)
-
-
-def fix_test():
-    tst = [
-        "spotify:track:6zrR8itT7IfAdl5aS7YQyt",
-        "spotify:track:60a0Rd6pjrkxjPbaKzXjfq",
-        "spotify:track:3gVhsZtseYtY1fMuyYq06F",
-        "spotify:track:5DXGHZ3QDh0FcLXPMWTv9U",
-        "spotify:track:60eOMEt3WNVX1m1jmApmnX"
-    ]
-    sp.playlist_remove_all_occurrences_of_items(
-        playlist_id=test,
-        items=tst,
-    )
-    sp.playlist_add_items(
-        playlist_id=test,
-        items=tst
-    )
-
+    sp_api.update_playlist_items(playlist_id=collection_id, uris=collection_tracks)
 
 
 if __name__ == '__main__':
-    test = url_to_uri("https://open.spotify.com/playlist/6bRkO7PLCXgmV4EJH52iU4?si=fbe6a250558f45f8", to_id=True)  # track_count=5
-    prod = url_to_uri("https://open.spotify.com/playlist/6QjbdNFUe4SFNE82RTmcCJ?si=ca49189f2c994135", to_id=True)  # track_count=98
-
-    # fix_test()
-    #
-    # # Test case: shuffle
-    # organize_playlist(
-    #     playlist_id=test,
-    #     shuffle=True
+    """"""
+    # organize_collection(
+    #     collection_uri="spotify:playlist:6bRkO7PLCXgmV4EJH52iU4",
+    #     shuffle=True,
+    #     remove=["duplicate"]
     # )
-    # # -> passed
-    #
-    # # Test case: remove
-    # organize_playlist(
-    #     playlist_id=test,
-    #     remove=[
-    #         "Live in",
-    #         "Live from",
-    #         "Acoustic",
-    #         "duplicate"
-    #     ]
-    # )
-    shuffle_lists = {
-    # "linkin_park_shuffle": url_to_uri("https://open.spotify.com/playlist/1TufW5MXE6oDxo7KVw4ACV?si=b923e75e5b3c4065"),
-    "current_shuffle": url_to_uri("https://open.spotify.com/playlist/6QjbdNFUe4SFNE82RTmcCJ?si=6df960bfc1f044b9"),
-    # "depr_shuffle": url_to_uri("https://open.spotify.com/playlist/5kdy1Iw5b2ZfYoxJxJIGxi?si=9b16d37cb21c4bcc")
-    }
-    for name,uri in shuffle_lists.items():
-        organize_collection(
-            collection_uri=uri,
-            shuffle=True,
-            remove=[
-                "duplicate"
-            ]
-        )
-        print(f"organized {name}")
-        sleep(2)  # <- no specific reason to be there
-
-    # image = image_to_b64(Image.open("/home/simon/git_repos/MusicPlayer/Icons/Spotipy_Logo.png"), 'PNG')
-    # if sp:
-    #     sp.playlist_upload_cover_image(playlist_id="1TufW5MXE6oDxo7KVw4ACV", image_b64=image)