+++ code_backend/testing/dev_bench_2.py
@@ -1,7 +1,8 @@
 """
 File to develop and debug methods, class and more
 """
+from code_backend.shared_config import *
 
 
-if __name__ == "__main__":
+if __name__ == '__main__':
     """"""