+++ code_backend/testing/dev_bench_3.py
@@ -1,6 +1,7 @@
 """
 File to develop and debug methods, class and more
 """
+from code_backend.shared_config import *
 
 
 if __name__ == '__main__':